added changes

+++ APLESSON_03/ex_02.java
@@ -8,13 +8,13 @@ public class ex_02
 		
 		System.out.println("What is your height? ");
 		
-		int height = keyboard.nextInt();
+		double height = keyboard.nextInt();
 		System.out.println( height + " Okay. What is your weight? ");
 		
-		int weight = keyboard.nextInt();
+		double weight = keyboard.nextInt();
 		System.out.println( weight + " Okay. Let me calculate your BMI. ");
 		
-		int BMI = 703 * (weight / height * height);
+		double BMI =  (703 *weight) / (height * height);
 		System.out.println( "Your BMI is " + BMI );
 		
 	}
